+++ Sources/CardPayments/Models/OrdersV2/PurchaseUnit.swift
@@ -1,4 +1,0 @@
-struct PurchaseUnit: Codable {
-
-    let amount: Amount
-}