+++ Sources/CardPayments/Models/ThreeDSecure/Attributes.swift
@@ -1,4 +1,0 @@
-struct Attributes: Codable {
-
-    let verification: Verification
-}