+++ Sources/CardPayments/Models/ThreeDSecure/Verification.swift
@@ -1,4 +1,0 @@
-struct Verification: Codable {
-
-    let method: String
-}