+++ Sources/PayPalNativePayments/PayPalNativeCheckoutResult.swift
@@ -6,6 +6,6 @@ public struct PayPalNativeCheckoutResult {
     /// The order ID associated with the transaction.
     public let orderID: String
 
-    /// The Payer ID (or user id) associated with the transaction.
+    /// The Payer ID (or user ID) associated with the transaction.
     public let payerID: String
 }