+++ Sources/CardPayments/Models/OrdersV2/Amount.swift
@@ -1,5 +1,0 @@
-struct Amount: Codable {
-
-    let currencyCode: String
-    let value: String
-}