+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 # PayPal iOS SDK Release Notes
 
+## unreleased
+* Update `PayPalNativeCheckoutDelegate.paypal(_:didFinishWithResult:)` to use `PayPalNativeCheckoutResult` instead of `PayPalCheckout.Approval` type.
+
 ## 0.0.6 (2023-02-21)
 * Fix CocoaPods build error for Xcode 13