+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Add `PayPalNativeShippingDelegate.didShippingAddressChange()`
     * Add `PayPalNativeShippingDelegate.didShippingMethodChange()`
     * Remove `PayPalNativeCheckoutDelegate.paypalDidShippingAddressChange()`
+  * Update `PayPalNativeCheckoutDelegate.paypal(_:didFinishWithResult:)` to use `PayPalNativeCheckoutResult` instead of `PayPalCheckout.Approval` type.
 
 ## 0.0.6 (2023-02-21)
 * Fix CocoaPods build error for Xcode 13